+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/keyvalue/api/watch/LockWatchEventCacheIntegrationTest.java
@@ -17,20 +17,28 @@
 package com.palantir.atlasdb.keyvalue.api.watch;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import java.nio.charset.Charset;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 
-import org.assertj.core.api.Assertions;
 import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.TestName;
 
-import com.codahale.metrics.MetricRegistry;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.datatype.guava.GuavaModule;
+import com.fasterxml.jackson.datatype.jdk8.Jdk8Module;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
-import com.palantir.atlasdb.util.MetricsManager;
 import com.palantir.lock.AtlasRowLockDescriptor;
 import com.palantir.lock.LockDescriptor;
 import com.palantir.lock.v2.LockToken;
@@ -46,7 +54,7 @@ import com.palantir.lock.watch.LockWatchStateUpdate;
 import com.palantir.lock.watch.TransactionUpdate;
 import com.palantir.lock.watch.TransactionsLockWatchUpdate;
 import com.palantir.lock.watch.UnlockEvent;
-import com.palantir.tritium.metrics.registry.DefaultTaggedMetricRegistry;
+import com.palantir.logsafe.exceptions.SafeIllegalArgumentException;
 
 public class LockWatchEventCacheIntegrationTest {
     private static final String TABLE = "table";
@@ -54,17 +62,19 @@ public class LockWatchEventCacheIntegrationTest {
     private static final LockDescriptor DESCRIPTOR_2 = AtlasRowLockDescriptor.of(TABLE, new byte[] {2});
     private static final LockDescriptor DESCRIPTOR_3 = AtlasRowLockDescriptor.of(TABLE, new byte[] {3});
     private static final LockWatchReferences.LockWatchReference REFERENCE = LockWatchReferences.entireTable("table");
-    private static final LockToken COMMIT_TOKEN = LockToken.of(UUID.randomUUID());
+    private static final UUID COMMIT_UUID = UUID.fromString("203fcd7a-b3d7-4c2a-9d2c-3d61cde1ba59");
+    private static final LockToken COMMIT_TOKEN = LockToken.of(COMMIT_UUID);
 
     private static final LockWatchEvent WATCH_EVENT =
-            LockWatchCreatedEvent.builder(ImmutableSet.of(REFERENCE), ImmutableSet.of(DESCRIPTOR_2)).build(4L);
+            LockWatchCreatedEvent.builder(ImmutableSet.of(REFERENCE), ImmutableSet.of(DESCRIPTOR)).build(4L);
     private static final LockWatchEvent UNLOCK_EVENT = UnlockEvent.builder(ImmutableSet.of(DESCRIPTOR_2)).build(5L);
     private static final LockWatchEvent LOCK_EVENT =
             LockEvent.builder(ImmutableSet.of(DESCRIPTOR_3), COMMIT_TOKEN).build(6L);
+    private static final UUID EVENT2_UUID = UUID.fromString("888fcd7a-b3d7-4d2a-9d2c-3d61cde1ba44");
     private static final LockWatchEvent LOCK_EVENT_2 =
-            LockEvent.builder(ImmutableSet.of(DESCRIPTOR), LockToken.of(UUID.randomUUID())).build(10L);
+            LockEvent.builder(ImmutableSet.of(DESCRIPTOR), LockToken.of(EVENT2_UUID)).build(10L);
 
-    private static final UUID LEADER = UUID.randomUUID();
+    private static final UUID LEADER = UUID.fromString("470c855e-f77b-44df-b56a-14d3df085dbc");
     private static final LockWatchStateUpdate SNAPSHOT =
             LockWatchStateUpdate.snapshot(LEADER, 3L, ImmutableSet.of(DESCRIPTOR_2), ImmutableSet.of());
     private static final LockWatchStateUpdate SUCCESS =
@@ -72,51 +82,162 @@ public class LockWatchEventCacheIntegrationTest {
     private static final long START_TS = 1L;
     private static final Set<TransactionUpdate> COMMIT_UPDATE = ImmutableSet.of(
             ImmutableTransactionUpdate.builder().startTs(START_TS).commitTs(5L).writesToken(COMMIT_TOKEN).build());
+    private static final Set<Long> TIMESTAMPS = ImmutableSet.of(START_TS);
+    private static final Set<Long> TIMESTAMPS_2 = ImmutableSet.of(16L);
+    private static final String BASE = "src/test/resources/lockwatch-event-cache-output/";
+    private static final Mode MODE = Mode.CI;
 
-    private final MetricsManager metricsManager = new MetricsManager(
-            new MetricRegistry(),
-            new DefaultTaggedMetricRegistry(),
-            unused -> false);
+    private enum Mode {
+        DEV,
+        CI;
+
+        boolean isDev() {
+            return this.equals(Mode.DEV);
+        }
+
+    }
 
     private LockWatchEventCache eventCache;
+    private int part;
+
+    @Rule
+    public TestName name = new TestName();
 
     @Before
     public void before() {
-        eventCache = LockWatchEventCacheImpl.create(metricsManager);
+        eventCache = new LockWatchEventCacheImpl(LockWatchEventLogImpl.create());
+        part = 1;
+    }
+
+    private void verifyStage() {
+        ObjectMapper mapper = new ObjectMapper()
+                .enable(SerializationFeature.INDENT_OUTPUT)
+                .registerModule(new Jdk8Module())
+                .registerModule(new GuavaModule());
+        try {
+            Path path = Paths.get(BASE + name.getMethodName() + "/event-cache-" + part + ".json");
+
+            if (MODE.isDev()) {
+                mapper.writeValue(path.toFile(), eventCache);
+            } else {
+                String ourJson = mapper.writeValueAsString(eventCache);
+                String theirJson = new String(Files.readAllBytes(path), Charset.defaultCharset());
+                assertThat(ourJson).isEqualTo(theirJson);
+            }
+            part++;
+        } catch (Throwable t) {
+            throw new RuntimeException(t);
+        }
+    }
+
+    @Test
+    public void processStartTimestampUpdateOnMultipleBatches() {
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        verifyStage();
+
+        Set<Long> secondTimestamps = ImmutableSet.of(5L, 123L);
+        eventCache.processStartTransactionsUpdate(secondTimestamps, SUCCESS);
+        verifyStage();
     }
 
     @Test
     public void getCommitUpdateDoesNotContainCommitLocks() {
-        Set<Long> timestamps = ImmutableSet.of(START_TS);
-        eventCache.processStartTransactionsUpdate(timestamps, SNAPSHOT);
-        assertThat(eventCache.lastKnownVersion()).hasValue(IdentifiedVersion.of(LEADER, 3L));
-
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
         eventCache.processGetCommitTimestampsUpdate(COMMIT_UPDATE, SUCCESS);
-        assertThat(eventCache.lastKnownVersion()).hasValue(IdentifiedVersion.of(LEADER, 6L));
+        verifyStage();
 
-        Assertions.assertThat(eventCache.getCommitUpdate(START_TS).accept(new CommitUpdateVisitor()))
-                .containsExactlyInAnyOrder(DESCRIPTOR_2);
+        CommitUpdate commitUpdate = eventCache.getCommitUpdate(1L);
+        assertThat(commitUpdate.accept(new CommitUpdateVisitor()))
+                .containsExactlyInAnyOrder(DESCRIPTOR);
+    }
+
+    @Test
+    public void cacheClearedOnSnapshotUpdate() {
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2, SUCCESS);
+        verifyStage();
+
+        LockWatchStateUpdate snapshot2 = LockWatchStateUpdate.snapshot(LEADER, 7L, ImmutableSet.of(DESCRIPTOR),
+                ImmutableSet.of());
+        Set<Long> timestamps3 = ImmutableSet.of(123L, 1255L);
+        eventCache.processStartTransactionsUpdate(timestamps3, snapshot2);
+        verifyStage();
     }
 
     @Test
     public void getEventsForTransactionsReturnsSnapshotWithOldEvents() {
-        Set<Long> timestamps = ImmutableSet.of(START_TS);
-        eventCache.processStartTransactionsUpdate(timestamps, SNAPSHOT);
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
         eventCache.processGetCommitTimestampsUpdate(COMMIT_UPDATE, SUCCESS);
         eventCache.removeTransactionStateFromCache(START_TS);
+        verifyStage();
 
         LockWatchStateUpdate success2 = LockWatchStateUpdate.success(LEADER, 10L, ImmutableList.of(LOCK_EVENT_2));
-        Set<Long> timestamps2 = ImmutableSet.of(16L);
-        eventCache.processStartTransactionsUpdate(timestamps2, success2);
-        assertThat(eventCache.lastKnownVersion()).hasValue(IdentifiedVersion.of(LEADER, 10L));
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2, success2);
+        verifyStage();
 
-        TransactionsLockWatchUpdate results = eventCache.getUpdateForTransactions(timestamps2, Optional.empty());
+        TransactionsLockWatchUpdate results = eventCache.getUpdateForTransactions(TIMESTAMPS_2, Optional.empty());
         assertThat(results.clearCache()).isTrue();
         assertThat(results.startTsToSequence()).containsExactlyInAnyOrderEntriesOf(
                 ImmutableMap.of(16L, IdentifiedVersion.of(LEADER, 10L)));
         assertThat(results.events()).containsExactly(
-                LockWatchCreatedEvent.builder(ImmutableSet.of(REFERENCE), ImmutableSet.of(DESCRIPTOR_3)).build(6L),
+                LockWatchCreatedEvent.builder(ImmutableSet.of(REFERENCE),
+                        ImmutableSet.of(DESCRIPTOR, DESCRIPTOR_3)).build(6L),
                 LOCK_EVENT_2);
+    }
+
+    @Test
+    public void failedUpdateClearsAllCaches() {
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2, LockWatchStateUpdate.failed(LEADER));
+        verifyStage();
+    }
+
+    @Test
+    public void leaderChangeClearsCaches() {
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2,
+                LockWatchStateUpdate.success(EVENT2_UUID, 4L, ImmutableList.of()));
+        verifyStage();
+    }
+
+    @Test
+    public void removingEntriesRetentionsEventsInLog() {
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2, SUCCESS);
+        verifyStage();
+
+        eventCache.removeTransactionStateFromCache(START_TS);
+        verifyStage();
+    }
+
+    @Test
+    public void nonContiguousEventsThrows() {
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2,
+                LockWatchStateUpdate.success(LEADER, 10L, ImmutableList.of(WATCH_EVENT, LOCK_EVENT, LOCK_EVENT_2)));
+        eventCache.processStartTransactionsUpdate(ImmutableSet.of(23L, 55L),
+                LockWatchStateUpdate.success(LEADER, 20L,
+                        ImmutableList.of(UnlockEvent.builder(ImmutableSet.of()).build(20L))));
+        verifyStage();
+        assertThatThrownBy(() -> eventCache.removeTransactionStateFromCache(START_TS))
+                .isExactlyInstanceOf(SafeIllegalArgumentException.class)
+                .hasMessage("Events form a non-contiguous sequence");
+    }
+
+    @Test
+    public void snapshotMissedEventThrows() {
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS, SNAPSHOT);
+        eventCache.processStartTransactionsUpdate(TIMESTAMPS_2,
+                LockWatchStateUpdate.success(LEADER, 5L, ImmutableList.of(UNLOCK_EVENT)));
+        eventCache.processStartTransactionsUpdate(ImmutableSet.of(23L, 55L),
+                LockWatchStateUpdate.success(LEADER, 6L, ImmutableList.of(LOCK_EVENT)));
+        verifyStage();
+
+        eventCache.removeTransactionStateFromCache(START_TS);
+        verifyStage();
+        assertThatThrownBy(() -> eventCache.removeTransactionStateFromCache(16L))
+                .isExactlyInstanceOf(SafeIllegalArgumentException.class)
+                .hasMessage("Events missing between last snapshot and this batch of events");
     }
 
     private static final class CommitUpdateVisitor implements CommitUpdate.Visitor<Set<LockDescriptor>> {
