+++ atlasdb-impl-shared/src/test/resources/lockwatch-event-cache-output/getEventsForTransactionsReturnsSnapshotWithOldEvents/event-cache-2.json
@@ -1,45 +1,22 @@
 {
   "eventLog" : {
     "snapshot" : {
-      "watches" : [ ],
+      "watches" : [ {
+        "type" : "fullTable",
+        "qualifiedTableRef" : "table"
+      } ],
       "locked" : [ {
-        "bytes" : "dGFibGUAAg=="
+        "bytes" : "dGFibGUAAQ=="
+      }, {
+        "bytes" : "dGFibGUAAw=="
       } ],
       "snapshotVersion" : {
         "id" : "470c855e-f77b-44df-b56a-14d3df085dbc",
-        "version" : 3
+        "version" : 6
       }
     },
     "eventStore" : {
       "eventMap" : {
-        "4" : {
-          "type" : "created",
-          "sequence" : 4,
-          "references" : [ {
-            "type" : "fullTable",
-            "qualifiedTableRef" : "table"
-          } ],
-          "lockDescriptors" : [ {
-            "bytes" : "dGFibGUAAQ=="
-          } ]
-        },
-        "5" : {
-          "type" : "unlock",
-          "sequence" : 5,
-          "lockDescriptors" : [ {
-            "bytes" : "dGFibGUAAg=="
-          } ]
-        },
-        "6" : {
-          "type" : "lock",
-          "sequence" : 6,
-          "lockDescriptors" : [ {
-            "bytes" : "dGFibGUAAw=="
-          } ],
-          "lockToken" : {
-            "requestId" : "203fcd7a-b3d7-4c2a-9d2c-3d61cde1ba59"
-          }
-        },
         "10" : {
           "type" : "lock",
           "sequence" : 10,
